requirements.txt
numpy~=2.3.3
matplotlib~=3.10.6
PyYAML~=6.0.3
scipy~=1.16.2